+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: `Tag`: Add `flex` prop, to behave as flexbox container.
+- Feature: `ProjectCard`: Add `topLine` prop, accepts React node.
+- Feature: `ProjectCard`: Add `Avatar` subcomponent, accepts `imageProps` prop.
+- Feature: `ProjectCard`: Add `noMargin` prop to `ItemsLine` subcomponent.
+
 ## [8.8.0] - 2022-03-08
 
 Deprecation/Change: